server/requirements.txt
flask
python-dotenv
requests

flask-sqlalchemy
flask-bcrypt
uuid
flask-session
redis
flask-cors

PyMySQL
cryptography
pydantic
Flask-Pydantic

pytest 
pytest-flask